Re-org, tensorboard, add gitignore
