--- projects/lsuv_weight_init/wi_experiments.cfg
@@ -1,42 +1,0 @@
-# ----------------------------------------------------------------------
-# Numenta Platform for Intelligent Computing (NuPIC)
-# Copyright (C) 2019, Numenta, Inc.  Unless you have an agreement
-# with Numenta, Inc., for a separate license for this software code, the
-# following terms and conditions apply:
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Affero Public License version 3 as
-# published by the Free Software Foundation.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-# See the GNU Affero Public License for more details.
-#
-# You should have received a copy of the GNU Affero Public License
-# along with this program.  If not, see http://www.gnu.org/licenses.
-#
-# http://numenta.org/licenses/
-# ----------------------------------------------------------------------
-
-[DEFAULT]
-path = results
-
-experiment = grid
-
-# Common network parameters
-iterations = 1 # 200
-test_batch_size = 64
-batch_size = 64
-batches_in_epoch = 400
-learning_rate = 0.1
-momentum = 0.9
-
-stop = {"stop": 1}
-checkpoint_at_end = True
-
-[LSUV]
-weight_init = lsuv
-
-[GRASS]
-weight_init = grassmanian
