+++ projects/weight_init/.gitignore
@@ -1,0 +1,2 @@
+data/
+runs/
