+++ projects/weight_init/results/scalars.json
@@ -1,0 +1,1 @@
+{"runs/Jun06_09-48-28_Jeremys-MacBook-Pro.local/%s/test/loss": [[1559839738.6086988, 0, 0.052456479024887086], [1559839769.379356, 1, 0.03601701233386993], [1559839799.235043, 2, 0.027762238621711732], [1559839828.7937531, 3, 0.030567320442199706], [1559839858.393473, 4, 0.03327054650783539], [1559839887.8596852, 5, 0.029638114070892334], [1559839917.2110112, 6, 0.024384280681610106], [1559839947.617672, 7, 0.030324587750434874], [1559839977.097526, 8, 0.03247716493606567], [1559840006.682786, 9, 0.0277455997467041], [1559840035.776653, 10, 0.026197078466415405], [1559840065.6187, 11, 0.027734506607055665], [1559840094.4312909, 12, 0.026524173069000245], [1559840123.212173, 13, 0.02761374785900116], [1559840152.022954, 14, 0.027981630325317385], [1559840180.769886, 15, 0.028299833512306215], [1559840210.121192, 16, 0.02856962413787842], [1559840239.247467, 17, 0.02894388518333435], [1559840268.33726, 18, 0.029104977750778197], [1559840297.543875, 19, 0.029496265268325807]], "runs/Jun06_09-48-28_Jeremys-MacBook-Pro.local/%s/test/correct": [[1559839738.6086988, 0, 9821.0], [1559839769.379356, 1, 9875.0], [1559839799.235043, 2, 9897.0], [1559839828.7937531, 3, 9895.0], [1559839858.393473, 4, 9900.0], [1559839887.8596852, 5, 9898.0], [1559839917.2110112, 6, 9921.0], [1559839947.617672, 7, 9906.0], [1559839977.097526, 8, 9912.0], [1559840006.682786, 9, 9922.0], [1559840035.776653, 10, 9929.0], [1559840065.6187, 11, 9930.0], [1559840094.4312909, 12, 9927.0], [1559840123.212173, 13, 9930.0], [1559840152.022954, 14, 9931.0], [1559840180.769886, 15, 9928.0], [1559840210.121192, 16, 9929.0], [1559840239.247467, 17, 9925.0], [1559840268.33726, 18, 9932.0], [1559840297.543875, 19, 9931.0]]}
